# docs/public/api/openapi-3.0.yaml
openapi: 3.0.0
info:
  description: Track, Manage, and Organize your Things.
  title: Homebox API
  contact:
    name: Homebox Team
    url: https://discord.homebox.software
  version: "1.0"
paths:
...
  /v1/users/change-password:
    put:
      security:
        - Bearer: []
      tags:
        - User
      summary: Change Password
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/v1.ChangePassword"
        description: Password Payload
        required: true
      responses:
        "204":
          description: No Content
...
  /v1/users/register:
    post:
      tags:
        - User
      summary: Register New User
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/services.UserRegistration"
        description: User Data
        required: true
      responses:
        "204":
          description: No Content
        "403":
          description: Local login is not enabled
          content:
            application/json:
              schema:
                type: string
...
  /v1/users/self:
    get:
      security:
        - Bearer: []
      tags:
        - User
      summary: Get User Self
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/v1.Wrapped"
                  - type: object
                    properties:
                      item:
                        $ref: "#/components/schemas/repo.UserOut"
    put:
      security:
        - Bearer: []
      tags:
        - User
      summary: Update Account
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/repo.UserUpdate"
        description: User Data
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/v1.Wrapped"
                  - type: object
                    properties:
                      item:
                        $ref: "#/components/schemas/repo.UserUpdate"
    delete:
      security:
        - Bearer: []
      tags:
        - User
      summary: Delete Account
      responses:
        "204":
          description: No Content
...
  /v1/users/self/settings:
    get:
      security:
        - Bearer: []
      tags:
        - User
      summary: Get user settings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/v1.Wrapped"
                  - type: object
                    properties:
                      item:
                        type: object
                        additionalProperties: true
    put:
      security:
        - Bearer: []
      tags:
        - User
      summary: Update user settings
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
        description: Settings Data
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/v1.Wrapped"
                  - type: object
                    properties:
                      item:
                        type: object
                        additionalProperties: true
servers:
  - url: /api
components:
  securitySchemes:
    Bearer:
      description: "\"Type 'Bearer TOKEN' to correctly set the API Key\""
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    repo.UserOut:
      type: object
      properties:
        defaultGroupId:
          type: string
        email:
          type: string
        groupIds:
          type: array
          items:
            type: string
        id:
          type: string
        isSuperuser:
          type: boolean
        name:
          type: string
        oidcIssuer:
          type: string
        oidcSubject:
          type: string
...
    repo.UserUpdate:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
...
    services.UserRegistration:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
        token:
          type: string
...
    v1.ChangePassword:
      type: object
      properties:
        current:
          type: string
        new:
          type: string
...
    v1.Wrapped:
      type: object
      properties:
        item: {}
...